refactor(web): extract use-interaction-state.ts from SpatialEditor
The gesture/selection interaction state (selectionState + applySelection,
gestureState + gestureStateRef, livePoint, snapGuides, marquee, the
canvasRef mirror, and the gesture-arming refs — doublePressRef,
lastPressRef, spaceDownRef, activePointerIdRef, navigationRef,
longPressRef + clearLongPress) moves verbatim into its own hook,
called first — `selectionState` is this component's earliest
declaration, so its call site precedes every other extracted hook's
even though this is the last cluster to move (entanglement order and
declaration order are independent, per the design).

Last of the six clusters. editor-state-surface.test.ts's sources glob
gains use-interaction-state.ts.

Verified against the un-refactored base (0ad749b4) in a scratch
worktree: shaped-node-editing.browser.test.tsx's two failures
("expected '     shapedbod' to contain 'shapedbody'") reproduce
identically there, so they predate this slice and are not a
regression from the move — a font-measurement environment issue, not
addressed here.
