+++ apps/web/src/components/spatial-editor/SpatialEditor.tsx
@@ -79,7 +79,7 @@ import {
 import { createBrowserMeasureText } from '@kamiazya/whiteboard-canvas-viewer'
 import type { CommentThread, SpatialCanvas, SpatialNode } from '@kamiazya/whiteboard-model'
 import { bundledFacetRegistry } from '@kamiazya/whiteboard-plugin-visual'
-import { forwardRef, type ReactNode, useImperativeHandle, useMemo, useRef, useState } from 'react'
+import { forwardRef, type ReactNode, useImperativeHandle, useMemo, useRef } from 'react'
 import { writeLastTool } from '@/lib/initial-tool'
 import type { ResolvedTheme } from '../../hooks/useThemeMode.js'
 import { parseClipboardText } from '../../lib/clipboard-fragment.js'
@@ -105,14 +105,7 @@ import { distanceToPolyline, findFreeSpot, hitTest, indexNodeBoxes } from './geo
 import { snapGesturePoint } from './gesture-snap.js'
 import { describeTarget, gestureTrace } from './gesture-trace.js'
 import { carriedByGesture } from './gesture-view.js'
-import type { GestureState } from './gestures.js'
-import {
-  createIdleState,
-  defaultCreateId,
-  NEW_NODE_HEIGHT,
-  NEW_NODE_WIDTH,
-  reduceGesture,
-} from './gestures.js'
+import { defaultCreateId, NEW_NODE_HEIGHT, NEW_NODE_WIDTH, reduceGesture } from './gestures.js'
 import { LinkEmbedLayer } from './LinkEmbedLayer.js'
 import { LinkUrlDialog } from './LinkUrlDialog.js'
 import { MarkdownNodeEditor } from './MarkdownNodeEditor.js'
@@ -124,7 +117,6 @@ import {
   DOUBLE_PRESS_WINDOW_MS,
   type NavigationEvent,
   type NavigationResult,
-  type NavigationState,
   type PointerKind,
   reduceNavigation,
 } from './navigation.js'
@@ -133,12 +125,7 @@ import { SelectionOverlay } from './SelectionOverlay.js'
 import { SnapGuidesOverlay } from './SnapGuidesOverlay.js'
 import { requiredTextNodeHeight } from './scene-render.js'
 import { keyedWithoutPrefix } from './scene-render-core.js'
-import {
-  EMPTY_SELECTION,
-  reduceSelection,
-  type SelectionEvent,
-  type SelectionState,
-} from './selection.js'
+import { reduceSelection } from './selection.js'
 import { isTextEntryEvent } from './shortcuts.js'
 import { TextNodeEditor } from './TextNodeEditor.js'
 import {
@@ -155,6 +142,7 @@ import { useEditSessionState } from './use-edit-session-state.js'
 import { useEditorKeyboard } from './use-editor-keyboard.js'
 import { MINIMAP_MIN_ROOT_WIDTH_PX, useEditorMeasurements } from './use-editor-measurements.js'
 import { EXPAND_MIN_H, EXPAND_MIN_W, useFileSeamScene } from './use-file-seam-scene.js'
+import { useInteractionState } from './use-interaction-state.js'
 import { useKeyboardAvoidance } from './use-keyboard-avoidance.js'
 import { useLockPolicy } from './use-lock-policy.js'
 import { useNativeCanvasListeners } from './use-native-canvas-listeners.js'
@@ -503,37 +491,29 @@ export const SpatialEditor = forwardRef<SpatialEditorHandle, SpatialEditorProps>
     const { shellRef, rootSize, inspectorIsSheet, viewportCenterScreen, containerSizeOf } =
       useEditorMeasurements(rootRef)
 
-    /**
-     * The multi-selection lives in ONE state object and every transition
-     * routes through the pure `reduceSelection` (selection.ts), so its
-     * invariants (primary never inside extras; extras only with a primary)
-     * hold by construction — never hand-write a primary/extras update pair.
-     * Functional updates make sequential events inside one handler compose
-     * instead of clobbering each other through stale closures.
-     */
-    const [selectionState, setSelectionState] = useState<SelectionState>(EMPTY_SELECTION)
+    const {
+      selectionState,
+      setSelectionState,
+      applySelection,
+      gestureState,
+      setGestureState,
+      gestureStateRef,
+      livePoint,
+      setLivePoint,
+      snapGuides,
+      setSnapGuides,
+      doublePressRef,
+      marquee,
+      setMarquee,
+      spaceDownRef,
+      lastPressRef,
+      activePointerIdRef,
+      navigationRef,
+      canvasRef,
+      longPressRef,
+      clearLongPress,
+    } = useInteractionState({ canvas })
     const selectedId = selectionState.primaryId
-    const applySelection = (event: SelectionEvent) =>
-      setSelectionState((prev) => reduceSelection(prev, event))
-    const [gestureState, setGestureState] = useState<GestureState>(createIdleState())
-    /**
-     * Live pointer position during an in-flight move/resize/connect, in canvas
-     * space. Component-local on purpose: the reducer still recomputes the real
-     * commit from startPoint at pointerup, so this drives ONLY the preview
-     * overlay below and never becomes a source of truth. Keeping it out of
-     * `canvas` is what stops a per-frame `renderCanvasToSvg` (measured at
-     * ~30ms on an 80-node canvas — far past a frame budget).
-     */
-    const [livePoint, setLivePoint] = useState<Point | null>(null)
-    /**
-     * Canvas-space lines justifying the current snap, cleared with the
-     * gesture. Same rationale as `livePoint`: a per-frame value that drives
-     * only an overlay, never the document.
-     */
-    const [snapGuides, setSnapGuides] = useState<{
-      readonly x: readonly number[]
-      readonly y: readonly number[]
-    } | null>(null)
     const {
       tool,
       setTool,
@@ -569,71 +549,6 @@ export const SpatialEditor = forwardRef<SpatialEditorHandle, SpatialEditorProps>
       viewportCenterScreen,
       containerSizeOf,
     })
-    /**
-     * Armed by a second same-target press inside the double-press window;
-     * RESOLVED at pointerup: zero movement opens the editor (node) or
-     * creates (empty), any movement means it was a drag all along. Firing
-     * at the release also sidesteps mousedown's default focus action, which
-     * used to blur the just-mounted textarea when we opened at the press.
-     */
-    const doublePressRef = useRef<{ key: string; point: Point } | null>(null)
-    /** In-flight marquee selection rect, in canvas space (Excalidraw
-     * semantics: plain drag on empty space selects; pan is Space+drag,
-     * middle-button drag, or wheel). */
-    const [marquee, setMarquee] = useState<{ start: Point; current: Point } | null>(null)
-    const spaceDownRef = useRef(false)
-    /**
-     * Last press for the SELECT tool's double press, keyed by what was under
-     * it — a node id, an edge id, or `'empty'`. Hand mode's own double press
-     * is not here: it has no target to key on, so the navigation machine
-     * holds it with the distance bound that keying cannot supply.
-     */
-    const lastPressRef = useRef<{ key: string; at: number; point: Point } | null>(null)
-    /**
-     * The pointerId this component currently holds capture for, or `null`.
-     * Tracked so unmount can best-effort release capture (see the teardown
-     * effect below) even though no window-level fallback listener exists to
-     * do it otherwise — mirrors `trySetPointerCapture`'s own
-     * best-effort/never-throw reasoning.
-     */
-    const activePointerIdRef = useRef<number | null>(null)
-    /**
-     * Everything navigation owns, as one value: which fingers are down, what
-     * is driving the viewport, what the last hand press was. See
-     * `navigation.ts` — the refs this replaced could not say when a gesture
-     * was over, and twice shipped a field that outlived one.
-     */
-    const navigationRef = useRef<NavigationState>(createIdleNavigation())
-
-    const canvasRef = useRef(canvas)
-    canvasRef.current = canvas
-
-    // Mirror for callbacks that outlive their render — the long-press timer
-    // fires ~500ms after the closure that armed it, by which time the press
-    // itself has usually advanced the gesture ('pressing'/'moving'); reducing
-    // a cancel against the ARM-time state would mis-apply it.
-    const gestureStateRef = useRef(gestureState)
-    gestureStateRef.current = gestureState
-
-    /**
-     * Touch long-press -> context menu. iOS Safari never synthesises a
-     * `contextmenu` event from a touch long-press (Android Chrome does), so
-     * without this the app's object menu is simply unreachable on an iPhone
-     * — the press reads as a drag start and the menu never opens. Armed on
-     * a single stationary touch, cancelled by movement past the slop, a
-     * second finger, or lift.
-     */
-    const longPressRef = useRef<{
-      timer: ReturnType<typeof setTimeout>
-      pointerId: number
-      screen: Point
-    } | null>(null)
-    const clearLongPress = () => {
-      if (longPressRef.current !== null) {
-        clearTimeout(longPressRef.current.timer)
-        longPressRef.current = null
-      }
-    }
     // The file-reference seam — the LOD gate, label/missing resolution and
     // the content cache — built once in useFileSeamScene and spread into
     // every scene-building call below (committed scene, drag ghost,
